+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow users to provide their own encryption passphrase for volumes
 - Allow users to provide their own encryption passphrase after cloning
+- Allow users to change the passphrase for a luks volume definition
 
 - ### Changed